Primeiras refatorações e mudança de arquivo
